backend/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
sentence-transformers
google-generativeai
langchain
pypdf
python-multipart
langchain-community
